Fix: user controller
- requesting user information is now with ID
- adjusted delete user route
- removed isLoggedIn field from sessiondata

+++ backend/src/controllers/user.controller.ts
@@ -20,12 +20,12 @@ export class UserController extends BaseController {
     }
 
     initializeRoutes(): void {
-		this.router.get('/:username', cors, this.requireAuth,
+		this.router.get('/:userid', cors, this.requireAuth,
 			upload.none(),
 			(req: express.Request, res: express.Response) => {
 				this.getUserInformation(req, res);
 			});
-		this.router.delete('/:username', cors, this.requireAuth,
+		this.router.delete('/:userid', cors, this.requireAuth,
 			upload.none(), this.requireAuth,
 			(req: express.Request, res: express.Response) => {
 				this.deleteUser(req, res);
@@ -44,16 +44,20 @@ export class UserController extends BaseController {
 
 	async deleteUser(req: express.Request, res: express.Response) {
 		const sessiondata = req.session;
-		await database.DeleteUser(sessiondata.userID);
-		sessiondata.isLoggedIn = false;
-		sessiondata.userID = null;
-		res.status(200).json({ success: true, message: "User successfully deleted"});
+		const loggedInUser = sessiondata.userID;
+		if (loggedInUser == req.params.userid) {
+			await database.DeleteUser(sessiondata.userID);
+			sessiondata.userID = null;
+			res.status(200).json({ success: true, message: "User successfully deleted"});
+		} else {
+			res.status(400).json({ success: false, error: "You do not have permission to delete this user!"});
+		}
 	}
 
 	async getUserInformation(req: express.Request, res: express.Response) {
 		const sessiondata = req.session;
-		const username = req.params.username;
-		const user = await database.RetrieveUser("username", username);
+		const userID = req.params.userid;
+		const user = await database.RetrieveUser("userID", userID);
 		if (user != null) {
 			const result = user.get({ plain: true});
 			delete result.password;
